==> scripts/05-ggplot2.Rmd <==
---
title: "Gráficos com ggplot2"
author: "Curso-R"
output: html_document
---

# Introdução

- Tese de doutorado do Hadley Wickham.

- Baseado na gramática dos gráficos (*grammar of graphics*) de Leland Wilkinson.

- Tenta responder a pergunta: o que é um gráfico estatístico?


> Um gráfico estatístico é uma representação visual dos dados que utiliza os atributos estéticos (posição, cor, forma, tamanho, ...) de formas geométricas (pontos, linhas, barras, ...).

- Com base nessa definição, Hadley desenvolveu a gramática dos gráficos em camadas (ou *layered grammar of graphics*): cada elemento do gráfico pode ser representado por uma camada e um gráfico seria a sobreposição dessas camadas.

```{r}
knitr::include_graphics("img/layers.png")
```


# Pacotes

```{r}
library(tidyverse)
library(ggplot2)
library(scales)
```

# Ler dados

```{r}
imdb <- read_rds("../dados/imdb.rds")
imdb <- imdb %>% mutate(lucro = receita - orcamento)

bd_exp <- read_rds("../dados/bd_exp.rds")
legendas <- read_rds("../dados/legendas.rds")
```

> Pergunta: será que filmes com maior orçamento tendem a arrecadar mais?

Gráfico de dispersão: arrecadação vs orçamento

```{r}
ggplot(imdb) +
  geom_point(mapping = aes(x = orcamento, y = receita))
```

Vamos acrescentar a reta y = x para visualizar a quantidade de filmes que não se pagaram.

```{r}
imdb %>%
  ggplot() +
  geom_point(mapping = aes(x = orcamento, y = receita)) +
  geom_abline(intercept = 0, slope = 1, color = "red")
```

Também podemos usar a cor dos pontos para visualizar os filmes que não se pagaram.

```{r}
imdb %>%
  mutate(
    lucro = ifelse(lucro <= 0, "Não", "Sim")
  ) %>%
  ggplot() +
  geom_point(mapping = aes(x = orcamento, y = receita, color = lucro))

# O que aconteceria se usássemos uma variável numérica para representar a cor dos pontos?
```


Vamos arrumar os labels.

```{r}
imdb %>%
  mutate(
    lucro = ifelse(lucro <= 0, "Não", "Sim")
  ) %>%
  filter(!is.na(lucro)) %>% 
  ggplot() +
  geom_point(mapping = aes(x = orcamento, y = receita, color = lucro)) +
  labs(x = "Orçamento", y = "Arrecadação", color = "Houve lucro?") 
```


Vamos também mudar as labels nos eixos.

```{r}
imdb %>%
  mutate(
    lucro = ifelse(lucro <= 0, "Não", "Sim")
  ) %>%
  filter(!is.na(lucro)) %>% 
  ggplot() +
  geom_point(mapping = aes(x = orcamento, y = receita, color = lucro)) +
  labs(x = "Orçamento", y = "Arrecadação", color = "Houve lucro?") +
  scale_y_continuous(label = scales::dollar_format()) +
  scale_x_continuous(label = scales::dollar_format()) +
  scale_color_manual(values = c("red", "green"), labels = c("N", "S"))
```

Também podemos colocar as variáveis na escala de milhões.

```{r}
imdb %>%
  mutate(
    lucro = ifelse(lucro <= 0, "Não", "Sim")
  ) %>% 
  mutate(
    receita = receita/1000000,
    orcamento = orcamento/1000000
  ) %>% 
  ggplot() +
  geom_point(mapping = aes(x = orcamento, y = receita, color = lucro)) +
  labs(
    x = "Orçamento  (em milhôes de dólares)", 
    y = "Arrecadação (em milhôes de dólares)", 
    color = "Houve lucro?"
  ) +
  scale_y_continuous(label = scales::dollar_format()) +
  scale_x_continuous(label = scales::dollar_format())
```


> Exercício: será que filmes mais caros tendem a receber uma nota maior?

1. Crie um gráfico de dispersão da nota do imdb pelo orçamento.

```{r}
bd_exp %>%
  ggplot(aes(x = sal_scmath_i35_1, y = sal_scmath_i38_1)) +
  geom_jitter(alpha = 0.1) +
  geom_smooth()
```

2. Pinte todos os pontos do gráfico de azul.

```{r}

```

3. Refaça o gráfico do item 1. apenas para filmes com classificação etária livre. Também refaça para os filmes com classificação etária "a partir de 13 anos".

```{r}

```

4. [bd_exp] Crie um gráfico de dispersão entre sal_scmath_i35_1 e sal_scmath_i38_1. Experimente usar geom_jitter() no lugar de geom_point() e veja a diferença.
```{r}

```



Podemos usar `facets` para fazer um gráfico para cada nível de uma variável.

```{r}
imdb %>%
  filter(classificacao %in% c("Livre", "A partir de 13 anos")) %>%
  ggplot() +
  geom_point(aes(x = orcamento, y = nota_imdb)) +
  facet_wrap(~classificacao)
```

A função `facet_grid()` permite combinar os níveis de duas variáveis.

```{r}
imdb %>%
  filter(classificacao %in% c("Livre", "A partir de 13 anos")) %>%
  mutate(ano_cat = ifelse(ano < 2000, "Antigos", "Recentes")) %>%
  mutate(ano_class = paste(classificacao, ano_cat)) %>% 
  ggplot() +
  geom_point(aes(x = orcamento, y = nota_imdb, color = ano_class)) +
  facet_grid(ano_cat ~ classificacao)
```


> Pergunta: quais são os diretores que fizeram mais filmes e quantos filmes eles fizeram?

Podemos fazer um gráfico de barras.

```{r}
imdb %>%
  ggplot() +
  geom_bar(aes(x = diretor))
```

Vamos visualizar apenas os 10 diretores com mais filmes.

```{r}
imdb %>% 
  dplyr::count(diretor) %>% 
  arrange(desc(n)) %>% 
  filter(!is.na(diretor)) %>% 
  dplyr::top_n(10, n) %>%
  ggplot() +
  geom_bar(aes(x = diretor, y = n), stat = "identity")

imdb %>% 
  dplyr::count(diretor) %>% 
  arrange(desc(n)) %>% 
  filter(!is.na(diretor)) %>% 
  dplyr::top_n(10, n) %>%
  ggplot() +
  geom_col(aes(x = diretor, y = n))
```

Vamos arrumar os labels rotacionando os nomes.

```{r}
imdb %>% 
  count(diretor) %>% 
  arrange(desc(n)) %>% 
  filter(!is.na(diretor)) %>% 
  top_n(10, n) %>%
  ggplot() +
  geom_bar(aes(x = diretor, y = n), stat = "identity") +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5))
```

Também seria interessante ordenar as barras de forma crescente. Para isso, precisamos lidar com `fatores`.

Um fator é a representação numérica de uma variavel categórica dentro do R. Ele será mostrado como texto (categoria/level), mas por trás ele é tratado como um inteiro.

O melhor pacote para tratar fatores é o `forcats`.

- `forcats`: `for` (para) + `cats` (categóricas)

- utilidade: funções para reordenar e renomear fatores de diversas formas

- é especialmente útil para visualização

- `fct_reorder`, `fct_infreq`, `fct_collapse`, `fct_lump`

- são usados dentro de mutate()

```{r}
library(forcats)

imdb %>% 
  count(diretor) %>% 
  arrange(desc(n)) %>% 
  filter(!is.na(diretor)) %>% 
  top_n(10, n) %>%
  mutate(
    diretor = fct_reorder(diretor, n)
  ) %>% 
  ggplot() +
  geom_bar(aes(x = diretor, y = n), stat = "identity") +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5))
```

Também podemos escolher outra cor para as barras.

```{r}
imdb %>% 
  count(diretor) %>% 
  arrange(desc(n)) %>% 
  filter(!is.na(diretor)) %>% 
  top_n(10, n) %>%
  mutate(
    diretor = fct_reorder(diretor, n)
  ) %>% 
  ggplot() +
  geom_bar(
    aes(x = diretor, y = n), 
    stat = "identity", 
    color = "darkblue",
    fill = "pink"
  ) +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5))

imdb %>% 
  count(diretor) %>% 
  arrange(desc(n)) %>% 
  filter(!is.na(diretor)) %>% 
  top_n(10, n) %>%
  mutate(
    diretor = fct_reorder(diretor, n)
  ) %>% 
  ggplot() +
  geom_bar(
    aes(x = diretor, y = n, fill = diretor), 
    stat = "identity", 
    color = "darkblue",
    show.legend = FALSE
  ) +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5))
```


# Boxplots

```{r}
imdb %>% 
  filter(!is.na(diretor)) %>%
  group_by(diretor) %>% 
  filter(n() >= 15) %>% 
  ggplot() +
  geom_boxplot(aes(x = diretor, y = lucro))
```

Ordenando pela mediana.

```{r}
imdb %>% 
  filter(!is.na(diretor)) %>%
  group_by(diretor) %>% 
  filter(n() >= 15) %>%
  ungroup() %>% 
  mutate(
    diretor = fct_reorder(diretor, lucro, .fun = max, na.rm = TRUE)
  ) %>% 
  ggplot() +
  geom_boxplot(aes(x = diretor, y = lucro, fill = diretor)) +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5))
```

Também poderíamos fazer um histograma para avaliar a distribuição.

```{r}
bd_exp %>% 
  ggplot() +
  geom_histogram(aes(x = Idade0))
```

Deixando mais bonito.

```{r}
bd_exp %>% 
  ggplot() +
  geom_histogram(
    aes(x = Idade0),
    color = "white",
    fill = "royalblue",
    binwidth = 0.5
  ) +
  theme_minimal()
```

> Exercício

1. Refaça o último gráfico utilizando a agora a variável `sal_cstrat_i16_1` como resposta.

```{r}
```


2.b Estude a distribuição da variável `sal_cstrat_i16_1` para cada turma.

```{r}

```

### Gráfico de barras

```{r}
bd_exp %>%
  count(sal_cexp_i03_1) %>%
  ggplot(aes(x = sal_cexp_i03_1, y = n)) +
  geom_col()
```

Exercício 1: faça um gráfico de barras para `sal_effper_i04_1` e **pinte as barras conforme a serie**.
Lembre-se de colocar `serie` dentro do `count()`!
Critique se o gráfico está bom.

```{r}

```

Exercício 2: copie o código do exercício 1 e refaça o gráfico utilizando o geom `geom_col(position = "dodge2")`

```{r}

```


## geom_line()

> Pergunta: qual o comportamento das notas dos filmes do Spielberg no IMDB ao longo do tempo?

Vamos fazer um gráfico de linhas das notas pelo ano.

```{r}
imdb %>% 
  filter(diretor == "Steven Spielberg") %>% 
  ggplot() +
  geom_line(aes(x = ano, y = nota_imdb))
```

Vamos agregar os dados para termos apenas uma medida por ano.

```{r}
imdb %>% 
  filter(diretor == "Steven Spielberg") %>%
  group_by(ano) %>% 
  summarise(nota_media = mean(nota_imdb, na.rm = TRUE)) %>% 
  ggplot() +
  geom_line(aes(x = ano, y = nota_media))
```

Podemos misturar dois `geom's`.

```{r}
imdb %>% 
  filter(diretor == "Steven Spielberg") %>%
  group_by(ano) %>% 
  summarise(nota_media = mean(nota_imdb, na.rm = TRUE)) %>% 
  ggplot() +
  geom_line(aes(x = ano, y = nota_media), alpha = 0.1) +
  geom_point(aes(x = ano, y = nota_media), alpha = 0.1) +
  geom_text(aes(x = ano, y = nota_media , label = nota_media))
  
```

Neste caso, é útil reescreveremos o código da seguinte maneira.

```{r}
imdb %>% 
  filter(diretor == "Steven Spielberg") %>%
  group_by(ano) %>% 
  summarise(nota_media = mean(nota_imdb, na.rm = TRUE)) %>% 
  ggplot(aes(x = ano, y = nota_media)) +
  geom_point(alpha = 0.1) +
  geom_line(alpha = 0.1) +
  geom_text(aes(label = nota_media))
```


> Exercícios

Repita as análises acima, agora para os atores (vamos começar utilizando apenas a coluna `ator_1`:

1. Selecione os atores com mais filmes. Visualize quem são os atores e o número de filmes que eles fizeram. Você pode ou não fazer a análise para um gênero.

```{r}

```


2. Selecione um dos atores e estude a nota de seus filmes no IMDB ao longo do tempo.

```{r}

```


(ATENÇÃO! Conteúdo avançado)
E se quiséssemos utilizar as 3 colunas de atores?

```{r}
imdb_atores <- imdb %>% 
  gather(posicao, ator, ator_1, ator_2, ator_3) %>% 
View(imdb_atores)
```

Repita a análise anterior, utilizando agora todas as colunas de atores, e veja se você mudaria de opinião sobre o ator a ser escolhido.

```{r}

```


# Misc de ggplot2

Temas.

```{r}
imdb %>%
  mutate(
    lucro = ifelse(lucro <= 0, "Não", "Sim")
  ) %>% 
  mutate(
    receita = receita/1000000,
    orcamento = orcamento/1000000
  ) %>% 
  ggplot() +
  geom_point(mapping = aes(x = orcamento, y = receita, color = lucro)) +
  labs(
    x = "Orçamento (em milhões de dólares)", 
    y = "Arrecadação (em milhões de dólares)", 
    color = "Houve lucro?"
  ) +
  scale_y_continuous(label = scales::dollar_format()) +
  scale_x_continuous(label = scales::dollar_format()) +
  theme_bw()
```

Títulos

```{r}
imdb %>%
  mutate(
    lucro = ifelse(lucro <= 0, "Não", "Sim")
  ) %>% 
  mutate(
    receita = receita/1000000,
    orcamento = orcamento/1000000
  ) %>% 
  ggplot() +
  geom_point(mapping = aes(x = orcamento, y = receita, color = lucro)) +
  labs(
    x = "Orçamento", 
    y = "Arrecadação", 
    color = "Houve lucro?"
  ) +
  scale_y_continuous(label = scales::dollar_format()) +
  scale_x_continuous(label = scales::dollar_format()) +
  theme_minimal() +
  ggtitle("Arrecadação vs Orçamento (em milhões de dólares)") +
  theme(plot.title = element_text(hjust = 0.5))
```

Legendas

```{r}
imdb %>%
  mutate(
    lucro = ifelse(lucro <= 0, "Não", "Sim")
  ) %>% 
  mutate(
    receita = receita/1000000,
    orcamento = orcamento/1000000
  ) %>% 
  ggplot() +
  geom_point(mapping = aes(x = orcamento, y = receita, color = lucro)) +
  labs(
    x = "Orçamento", 
    y = "Arrecadação", 
    color = "Houve lucro?"
  ) +
  scale_y_continuous(label = scales::dollar_format()) +
  scale_x_continuous(label = scales::dollar_format()) +
  theme_minimal() +
  ggtitle("Arrecadação vs Orçamento (em milhões de dólares)") +
  theme(legend.position = "bottom")
```

gplot2 extensions: https://www.ggplot2-exts.org/

```{r}
library(ggridges)

imdb %>%
  filter(ano > 2005) %>% 
  mutate(ano = as.factor(ano)) %>% 
  ggplot(aes(y = ano, x = receita, fill = ano)) +
  geom_density_ridges(na.rm = TRUE, show.legend = FALSE)
```

Foco

```{r}
imdb %>%
  filter(ano > 2005) %>% 
  mutate(ano = as.factor(ano)) %>% 
  ggplot(aes(y = ano, x = receita, fill = ano)) +
  geom_density_ridges(na.rm = TRUE) +
  coord_cartesian(xlim = c(-30000000, 200000000))
```



## Exercícios

1. Faça um gráfico para visualizar as 10 maiores bilheterias de 2016.

```{r}

```


2. Quantos filmes "são feitos por ano"?

```{r}

```


## Outros pacotes interessantes para gráficos:

Plotly

```{r}
library(plotly)

p <- imdb %>% 
  filter(diretor == "Steven Spielberg") %>%
  ggplot(aes(x = ano, y = nota_imdb, text = titulo)) +
  geom_point() +
  geom_line(aes(group = 1))

ggplotly(p, tooltip = "text")
```

Patchwork

```{r}
devtools::install_github("thomasp85/patchwork")

library(patchwork)

p1 <- imdb %>% 
  filter(!is.na(diretor)) %>%
  group_by(diretor) %>% 
  filter(n() >= 15) %>%
    filter(n() >= 15) %>%
  ungroup() %>% 
  mutate(
    diretor = as.factor(diretor),
    diretor = fct_reorder(diretor, lucro, .fun = mean, na.rm = TRUE)
  ) %>% 
  ggplot() +
  geom_boxplot(aes(x = diretor, y = lucro)) +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5))

p2 <- imdb %>% 
  filter(diretor == "Steven Spielberg") %>% 
  ggplot() +
  geom_histogram(
    aes(x = lucro),
    fill = "lightblue", 
    color = "darkblue", 
    binwidth = 100000000
  )


p1 + p2
```
