==> scripts/06-case-respostas.Rmd <==
---
title: "Case"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
base_ias <- read_rds("../dados/bd_exp.rds") %>% 
  janitor::clean_names()
```


### Tabela 1: notas de português e matemática

```{r}
base_ias %>%
  gather(materia, nota, a_m_port, a_m_mat) %>% 
  group_by(turma, materia, serie) %>% 
  summarise(
    nota_media = mean(nota, na.rm = TRUE)
  ) %>% 
  ungroup() %>% 
  mutate(nota_media = round(nota_media, 1)) %>% 
  arrange(materia, desc(nota_media)) %>% 
  mutate(
    materia = ifelse(materia == "a_m_mat", "Matemática", "Português")
  ) %>% 
  select(
    Turma = turma,
    Série = serie,
    Matéria = materia,
    `Nota média` = nota_media
  )
```

### Tabela 2: descritiva da nota geral ((PORT + MAT)/ 2) por turma do quinto ano

```{r}
base_ias %>% 
  filter(serie == 5) %>% 
  mutate(nota_geral = (a_m_mat + a_m_port) / 2) %>% 
  group_by(turma) %>% 
  summarise(
    nota_geral_media = mean(nota_geral, na.rm = TRUE),
    nota_geral_dp = sd(nota_geral, na.rm = TRUE),
    nota_geral_min = min(nota_geral, na.rm = TRUE),
    nota_geral_max = max(nota_geral, na.rm = TRUE),
  ) %>% 
  select(
    Turma = turma,
    `Média` = nota_geral_media,
    `Desvio-padrão` = nota_geral_dp,
    `Mínima` = nota_geral_min,
    `Máxima` = nota_geral_max
  ) %>% 
  arrange(desc(Média))
```


### Tabela 3: escore médio dos itens

```{r}
base_dic <- readxl::read_excel("../dados/senna_sal_ex1.xlsx", sheet = 2) %>% 
  mutate(
    coditem2 = str_to_lower(coditem2),
    coditem2 = str_replace(coditem2, "intcur", "int_cur"),
    coditem2 = str_replace(coditem2, "crimg", "cr_img"),
    coditem2 = str_replace(coditem2, "_n_l", "_n_l_"),
    coditem2 = str_replace(coditem2, "angrvol", "angr_vol")
  )

base_ias %>% 
  select(starts_with("sal"),  starts_with("senna")) %>% 
  gather(item, score) %>% 
  group_by(item) %>% 
  summarise(score_medio = mean(score, na.rm = TRUE)) %>% 
  arrange(desc(score_medio)) %>% 
  mutate(score_medio = round(score_medio, 1)) %>% 
  left_join(base_dic, by = c("item" = "coditem2")) %>% 
  select(
    Cod = item,
    Item = item_text,
    `Score médio` = score_medio
  )
```


### Gráfico 1: grafico_notas.png

```{r}
base_ias %>% 
  ggplot(aes(x = a_m_port, y = a_m_mat)) +
  geom_point() +
  geom_hline(yintercept = 60, linetype = 2) +
  geom_vline(xintercept = 60, linetype = 2) +
  geom_smooth(se = FALSE) +
  theme_minimal() +
  labs(x = "Nota de português", y = "Nota de matemática")
```


### Gráfico 2: grafico_escores_dominio_O.png

```{r}
base_ias %>% 
  select(starts_with("sal"),  starts_with("senna")) %>% 
  gather(item, score) %>%
  left_join(base_dic, by = c("item" = "coditem2")) %>% 
  filter(domain == "O") %>% 
  ggplot(aes(x = item, y = score)) +
  geom_boxplot() +
  coord_flip() +
  theme_minimal() +
  labs(y = "Escore", x = "")
```

### Gráfico 3

```{r}
library(scales)

tabela_sal_cexp_i03_1 <- base_ias %>%
  count(sal_cexp_i03_1) %>%
  mutate(
    p = n/sum(n),
    p_lbl = percent(p, 0.1, decimal.mark = ","),
    p_lbl = paste0(n, " (", p_lbl, ")")
  )

tabela_sal_cexp_i03_1

# Gráfico feio
grafico_sal_cexp_i03_1 <- tabela_sal_cexp_i03_1 %>%
  ggplot(aes(x = "sal_cexp_i03_1", y = p, fill = sal_cexp_i03_1)) +
  geom_col(position = "stack", width = 0.4) +
  geom_label(
    aes(label = p_lbl),
    position = position_stack(vjust = .5),
    colour = "white",
    fontface = "bold",
    label.size = 0
  )

grafico_sal_cexp_i03_1

# Deixando mais bonito
grafico_sal_cexp_i03_1 +
  coord_flip() +
  theme_minimal() +
  theme(legend.position = "top",
        axis.text.y = element_text(hjust = 0)
  ) +
  labs(fill = "Resposta:", x = "", y = "") +
  scale_y_continuous(labels = scales::percent_format(decimal.mark = ",")) +
  ggtitle("sal_cexp_i03_1") +
  scale_fill_gradient2(high = "#CA0020",  mid = "#dddddd", low = "#0571B0", midpoint = 2.5) 
ggsave("grafico_barras_rebuscado.png")
```
